Add node init
Instead of just having an update function, allow actions to define an
object with "init" and "update" function.

The init function is called if defined and the last running node of the
instance was not the same node.

We validate the node names to not be one of the reserved names.

We introduce a pass-through "ctx" object that can be used in "init" and
"update" to connect to the rest of the application.

We reorganize the test so that the individual tests are independent of
each other and each defines the actions it needs for the tested
behavior.