Refactor API.
For our purposes it seems best to not rely on individually imported
modules but a central API point for all purposes.

Behavior Facade
---------------

We make Behavior.js export a API facade object which offers the central
entry points of functionality.

 * load(data)

   load a behavior tree from the data-representation. Internally we just
   do a new Behavior(data). That constructor is also exported as
   facade.Behavior

 * registerAction(name,handler)

   Register a tree leaf handler object.

 * registerFromRequireContext(ctx)

   Register all actions from a webpack require.context() context. Is
   included for ease-of-use from webpack, requiring no dependencies etc.

   Should be easy to build something similar for "bulkify" or similar.

 * resetActions()

   Forget all action registrations.

The facadae also contains the Behavior and BehaviorInstance constructors
and the State enum as properties.